requirements.txt
asgiref==3.7.2
diff-match-patch==20241021
Django==4.2
django-import-export==4.3.3
fontawesomefree==6.5.1
Pillow==10.1.0
sqlparse==0.5.0
tablib==3.7.0
typing_extensions==4.9.0
tzdata==2024.2
